feat: implement GithubService for handling comments and integrate with WebhookAction
